Checked in changes to make peer compile again on Windows (32-bit, MinGW). Does not run succesfully, yet.
git-svn-id: svn+ssh://svn.fcdonders.nl/home/svnroot/fieldtrip/trunk@2244 0cf7c7f0-3615-4144-b4e6-68da3bce3cd0

+++ peer/src/announce.c
@@ -39,7 +39,7 @@
 
 float frand(float min, float max) {
 		float val;
-		val = rand();				/* ranging from 0 to RAND_MAX */
+		val = (float) rand();		/* ranging from 0 to RAND_MAX */
 		val = val/RAND_MAX;			/* ranging from 0 to 1 */
 		val = min + val*(max-min);	/* ranging from min to max */
 		return val;
